+++ ha-business/src/main/java/jp/co/ha/business/db/crud/read/impl/AccountRecoveryTokenSearchServiceImpl.java
@@ -30,7 +30,7 @@ public class AccountRecoveryTokenSearchServiceImpl
 
     @Override
     public Optional<AccountRecoveryTokenData> findBySeqUserIdAndTokenAndValidTokenCreateDate(
-            Integer seqUserId, String token) {
+            Long seqUserId, String token) {
 
         AccountRecoveryTokenDataExample example = new AccountRecoveryTokenDataExample();
         Criteria creteria = example.createCriteria();
